refactor : reciepient URL 로직 분리
